% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplet_formation.R
\name{duplet_formation}
\alias{duplet_formation}
\title{Build duplets based on mutations}
\usage{
duplet_formation(data, mutations)
}
\arguments{
\item{data}{A data frame containing mutations that will be classified}

\item{mutations}{Specify the column name that contains the mutations}
}
\description{
This function takes the data frame and mutation column to form WT-MT duplets. It also
prints the missing duplets within the dataset.
}
\examples{
duplet_formation(data = df, mutations = "Mutation")

}
