% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplet_reduction.R
\name{duplet_reduction}
\alias{duplet_reduction}
\title{Compilation of a non-redundant dataset based on duplets}
\usage{
duplet_reduction(data, duplets, DDG, SS = NULL, ASA = NULL)
}
\arguments{
\item{data}{A redundant dataset of mutations}

\item{duplets}{Column name of Duplets}

\item{DDG}{Column name of Experimental DDG values}

\item{SS}{Default is Null. Specify the column name that contains secondary structure information}

\item{ASA}{Default is Null. Specify the column name that contains relative ASA information}
}
\value{
A list of reduced dataset and a box plot
}
\description{
This function takes the data frame, labels, experimental DDG, and the mutations' biochemical
features. With this function, the dataset can be filtered based on duplets, secondary structure,
and relative ASA of the mutations.
}
\examples{
duplet_reduction(data = df, duplets = "Duplets", DDG = "Experimental")
duplet_reduction(data = df, duplets = "Duplets", DDG = "Experimental", SS = "SStructure")

}
