% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gror_data.R
\docType{data}
\name{gror_data}
\alias{gror_data}
\title{Example data to compile}
\format{
A list object
}
\usage{
data(gror_data)
}
\description{
The dataset contains experimental DD, secondary structure and relative ASA
features of the data points.
}
\examples{
data(gror_data)
class_formation(data=gror_data, mutations="Mutation")
duplet_formation(data=gror_data, mutations="Mutation")

}
\keyword{datasets}
