% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranged_class_reduction.R
\name{ranged_class_reduction}
\alias{ranged_class_reduction}
\title{Compilation of a non-redundant dataset based on classes}
\usage{
ranged_class_reduction(data, classes, DDG, r, SS = NULL, ASA = NULL)
}
\arguments{
\item{data}{A redundant dataset of mutations}

\item{classes}{Column name of labels}

\item{DDG}{Column name of Experimental DDG values}

\item{r}{difference between each interval}

\item{SS}{Default is Null. Specify the column name that contains secondary structure information}

\item{ASA}{Default is Null. Specify the column name that contains relative ASA information}
}
\value{
A list of reduced dataset and a box plot
}
\description{
This function filters the dataset based on class labels,
secondary structure, and relative ASA of the mutations. This function takes
four inputs: the data frame to reduce, class labels, experimental DDG values,
and the specified difference between intervals in kcal/mol. Secondary structure
and relative ASA of the mutations are also applicable for reduction.
}
\examples{
ranged_class_reduction(data = df, classes = "Classes",
                       DDG = "Experimental", r = 2)
ranged_class_reduction(data = df, classes = "Classes",
                       DDG = "Experimental", r = 2, SS = "SStructure")
ranged_class_reduction(data = df, classes = "Classes",
                       DDG = "Experimental", r = 2, ASA = "ASA")

}
