% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranged_duplet_reduction.R
\name{ranged_duplet_reduction}
\alias{ranged_duplet_reduction}
\title{Compilation of a non-redundant dataset based on duplets}
\usage{
ranged_duplet_reduction(data, duplets, DDG, r, SS = NULL, ASA = NULL)
}
\arguments{
\item{data}{A redundant dataset of mutations}

\item{duplets}{Column name of Duplets}

\item{DDG}{Column name of Experimental DDG values}

\item{r}{difference between each interval}

\item{SS}{Default is Null. Specify the column name that contains secondary structure information}

\item{ASA}{Default is Null. Specify the column name that contains relative ASA information}
}
\value{
A list of reduced dataset and a box plot
}
\description{
This function filters the dataset based on duplet labels,
secondary structure, and relative ASA of the mutations. This function takes
four inputs: the data frame to reduce, duplet labels, experimental DDG values,
and the specified difference between intervals in kcal/mol. Secondary structure
and relative ASA of the mutations are also applicable for reduction.
}
\examples{
ranged_class_reduction(data = df, duplets = "Duplets",
                       DDG = "Experimental", r = 2)
ranged_class_reduction(data = df, duplets = "Duplets",
                       DDG = "Experimental", r = 2, SS = "SStructure")
ranged_class_reduction(data = df, duplets = "Duplets",
                       DDG = "Experimental", r = 2, ASA = "ASA")

}
