% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt2mt.R
\name{wt2mt}
\alias{wt2mt}
\title{Visualize the distribution of the mutations}
\usage{
wt2mt(data, mutations)
}
\arguments{
\item{data}{A data frame}

\item{mutations}{Specify the column name that contains the mutations}
}
\description{
This function takes the data frame and the mutation column to visualize the frequency of the
mutations.
}
\examples{
wt2mt(data = df, mutations = "Mutation")

}
