% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_formation.R
\name{class_formation}
\alias{class_formation}
\title{Build class labels based on mutations}
\usage{
class_formation(data, mutations)
}
\arguments{
\item{data}{A data frame containing mutations that will be classified}

\item{mutations}{Specify the column name that contains the mutations}
}
\description{
This function takes the data frame and the mutation column to group the
amino acids into four main classes based on their R groups' properties. It also
prints the missing classes within the dataset.
}
\examples{
class_formation(data = df, mutations = "Mutation")

}
