% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gror_data.R
\docType{data}
\name{gror_data}
\alias{gror_data}
\title{Example data to compile}
\format{
A list object
}
\usage{
data(gror_data)
}
\description{
The dataset contains class & duplet labels, secondary structure and relative ASA
of the mutations.
}
\examples{
data(gror_data)
class_reduction(gror_data, "Classes", "Experimental")
duplet_reduction(gror_data, "Duplets", "Experimental")

}
\keyword{datasets}
